+++ packages/components/fields/date-field/src/date-field.spec.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { render, fireEvent } from '../../../../test/test-utils';
+import { render, fireEvent } from '../../../../../test/test-utils';
 import DateField from './date-field';
 
 // This component is used to enable easy testing.
