+++ packages/components/fields/date-field/src/date-field.story.js
@@ -10,9 +10,9 @@ import {
   object,
 } from '@storybook/addon-knobs/react';
 import { getExampleDateStrings } from '@commercetools-uikit/calendar-utils';
-import Section from '../../../../.storybook/decorators/section';
-import Readme from './README.md';
-import * as icons from '../../icons';
+import Section from '../../../../../.storybook/decorators/section';
+import Readme from '../README.md';
+import * as icons from '../../../icons';
 import DateField from './date-field';
 
 const exampleDates = getExampleDateStrings();
