+++ packages/components/fields/date-field/src/index.js
@@ -1,0 +1,3 @@
+export { default } from './date-field';
+
+export { default as version } from './version';
