--- src/components/fields/date-field/index.js
@@ -1,1 +1,0 @@
-export { default } from './date-field';
